synchronize type of RFID in the DB table

+++ models/BikeData.js
@@ -5,7 +5,7 @@ var BikeData = sql.define("BikeData", {
         stamp: {type: Sql.BIGINT, primaryKey: true},
         rpm: {type: Sql.FLOAT},
         bikeID: {type: Sql.INTEGER},
-        sessionID: {type: Sql.INTEGER}
+        sessionID: {type: Sql.BIGINT}
 }, {timestamps: false, freezeTableName: true});
     return BikeData;
 };
